Authentication and Authorization using Jwt

+++ controllers/api/v1/posts_api.js
@@ -22,9 +22,9 @@ module.exports.destroy = async (req, res) => {
     try {
       const post = await Post.findById(req.params.id);
       if (!post) {
-        return res.status(404).send("Post not found");
+        return res.status(404).send({message: "Post not found"});
       }
-    //   if (post.user.toString() === req.user.id) {
+      if (post.user.toString() === req.user.id) {
          post.deleteOne();
   
         // Delete comments associated with the post
@@ -33,10 +33,9 @@ module.exports.destroy = async (req, res) => {
             message: 'Post and associated comments are deleted successfully!'
         });
 
-    //   } else {
-    //     req.flash('error', 'You can not delete this post');
-    //     return res.status(403).send("You're not authorized to delete this post");
-    //   }
+      } else {
+        return res.status(401).send({message: "You're not authorized to delete this post"});
+      }
     } catch (err) {
       console.error(err);
       return res.status(500).send({error: err.message});
